ai-brain/requirements.txt
# ─── Core AI / Agent Framework ───────────────────────────────────────────────
langgraph>=0.1.0
langchain>=0.2.0
langchain-community>=0.2.0

# ─── LLM Providers ────────────────────────────────────────────────────────────
langchain-aws>=0.1.0          # AWS Bedrock (Claude 3.5 Sonnet — heavy reasoning)
ollama>=0.2.0                 # Ollama HTTP client (Qwen2.5-Coder — light tasks)

# ─── gRPC (communicates signals to Go backend) ────────────────────────────────
grpcio>=1.63.0
grpcio-tools>=1.63.0
protobuf>=4.25.0

# ─── Market Data (Yahoo Finance — NO API KEY REQUIRED) ────────────────────────
yfinance>=0.2.40              # Free Yahoo Finance data via public endpoints
pandas>=2.2.0                 # DataFrame operations for indicator computation

# ─── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
lxml>=5.0.0                   # HTML parsing for S&P 500 symbol universe fetch
pydantic>=2.7.0
httpx>=0.27.0                 # Async HTTP (polls Go backend for mode)
tenacity>=8.3.0               # Retry logic for LLM calls
structlog>=24.1.0             # Structured logging
